ERP-972 - Implementation Open Id Connect to ethos-docmosis-backend-se…

+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/BulkActionsController.java
@@ -23,6 +23,7 @@ import uk.gov.hmcts.ethos.replacement.docmosis.service.*;
 
 import java.util.List;
 
+import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static uk.gov.hmcts.ethos.replacement.docmosis.model.helper.Constants.*;
 
@@ -38,16 +39,18 @@ public class BulkActionsController {
     private final BulkSearchService bulkSearchService;
     private final DocumentGenerationService documentGenerationService;
     private final SubMultipleService subMultipleService;
+    private final VerifyTokenService verifyTokenService;
 
     @Autowired
     public BulkActionsController(BulkCreationService bulkCreationService, BulkUpdateService bulkUpdateService,
                                  BulkSearchService bulkSearchService, DocumentGenerationService documentGenerationService,
-                                 SubMultipleService subMultipleService) {
+                                 SubMultipleService subMultipleService, VerifyTokenService verifyTokenService) {
         this.bulkCreationService = bulkCreationService;
         this.bulkUpdateService = bulkUpdateService;
         this.bulkSearchService = bulkSearchService;
         this.documentGenerationService = documentGenerationService;
         this.subMultipleService = subMultipleService;
+        this.verifyTokenService = verifyTokenService;
     }
 
     @PostMapping(value = "/createBulk", consumes = APPLICATION_JSON_VALUE)
@@ -62,6 +65,11 @@ public class BulkActionsController {
             @RequestBody BulkRequest bulkRequest,
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("CREATE BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkCasesPayload bulkCasesPayload = bulkSearchService.bulkCasesRetrievalRequest(bulkRequest.getCaseDetails(), userToken, true);
 
@@ -85,6 +93,11 @@ public class BulkActionsController {
             @RequestBody BulkRequest bulkRequest,
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("CREATE BULKES ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         if (bulkRequest.getCaseDetails().getCaseData().getCaseSource() == null
                 || bulkRequest.getCaseDetails().getCaseData().getCaseSource().trim().equals("")) {
@@ -115,6 +128,11 @@ public class BulkActionsController {
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("AFTER SUBMITTED BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
 
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
         BulkCasesPayload bulkCasesPayload = bulkSearchService.bulkCasesRetrievalRequestElasticSearch(
                 bulkRequest.getCaseDetails(), userToken, true, false);
         //BulkCasesPayload bulkCasesPayload = bulkSearchService.bulkCasesRetrievalRequest(bulkRequest.getCaseDetails(), userToken, false);
@@ -140,6 +158,11 @@ public class BulkActionsController {
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("UPDATE BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
 
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
         BulkRequestPayload bulkRequestPayload = bulkUpdateService.bulkUpdateLogic(bulkRequest.getCaseDetails(), userToken);
 
         bulkRequestPayload = bulkUpdateService.clearUpFields(bulkRequestPayload);
@@ -163,6 +186,11 @@ public class BulkActionsController {
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("UPDATE BULK CASE IDS ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
 
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
         BulkRequestPayload bulkRequestPayload = bulkCreationService.bulkUpdateCaseIdsLogic(bulkRequest, userToken);
 
         return ResponseEntity.ok(BulkCallbackResponse.builder()
@@ -183,6 +211,11 @@ public class BulkActionsController {
             @RequestBody BulkRequest bulkRequest,
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("GENERATE BULK LETTER ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkDocumentInfo bulkDocumentInfo = documentGenerationService.processBulkDocumentRequest(bulkRequest, userToken);
         bulkRequest.getCaseDetails().getCaseData().setDocMarkUp(bulkDocumentInfo.getMarkUps());
@@ -206,6 +239,11 @@ public class BulkActionsController {
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("GENERATE BULK LETTER CONFIRMATION ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
 
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
         return ResponseEntity.ok(BulkCallbackResponse.builder()
                 .data(bulkRequest.getCaseDetails().getCaseData())
                 .confirmation_header(GENERATED_DOCUMENTS_URL + bulkRequest.getCaseDetails().getCaseData().getDocMarkUp())
@@ -221,8 +259,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> midSearchBulk(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("MID SEARCH BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = bulkSearchService.bulkMidSearchLogic(bulkRequest.getCaseDetails(), false);
 
@@ -241,8 +285,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> searchBulk(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("SEARCH BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = bulkSearchService.bulkSearchLogic(bulkRequest.getCaseDetails());
 
@@ -261,8 +311,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> midCreateSubMultiple(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("MID CREATE SUB MULTIPLE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = bulkSearchService.bulkMidSearchLogic(bulkRequest.getCaseDetails(), true);
 
@@ -281,8 +337,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> createSubMultiple(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("CREATE SUB MULTIPLE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.createSubMultipleLogic(bulkRequest.getCaseDetails());
 
@@ -301,8 +363,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> subMultipleDynamicList(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("SUB MULTIPLE DYNAMIC LIST ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.populateSubMultipleDynamicListLogic(bulkRequest.getCaseDetails());
 
@@ -321,8 +389,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> filterDefaultedAllDynamicList(
-            @RequestBody BulkRequest bulkRequest) {
-        log.info("FILTER DEFAULTED DYNAMIC LIST ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
+        log.info("FILTER DEFAULTED ALL DYNAMIC LIST ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.populateFilterDefaultedDynamicListLogic(bulkRequest.getCaseDetails(), SELECT_ALL_VALUE);
 
@@ -341,8 +415,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> filterDefaultedNoneDynamicList(
-            @RequestBody BulkRequest bulkRequest) {
-        log.info("FILTER DEFAULTED DYNAMIC LIST ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
+        log.info("FILTER DEFAULTED NONE DYNAMIC LIST ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.populateFilterDefaultedDynamicListLogic(bulkRequest.getCaseDetails(), SELECT_NONE_VALUE);
 
@@ -361,8 +441,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> midUpdateSubMultiple(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("MID UPDATE SUB MULTIPLE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.bulkMidUpdateLogic(bulkRequest.getCaseDetails());
 
@@ -381,8 +467,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> updateSubMultiple(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("UPDATE SUB MULTIPLE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.updateSubMultipleLogic(bulkRequest.getCaseDetails());
 
@@ -401,8 +493,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> deleteSubMultiple(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("DELETE SUB MULTIPLE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkRequestPayload bulkRequestPayload = subMultipleService.deleteSubMultipleLogic(bulkRequest.getCaseDetails());
 
@@ -424,6 +522,11 @@ public class BulkActionsController {
             @RequestBody BulkRequest bulkRequest,
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("GENERATE BULK SCHEDULE ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         BulkDocumentInfo bulkDocumentInfo = documentGenerationService.processBulkScheduleRequest(bulkRequest, userToken);
 
@@ -451,8 +554,14 @@ public class BulkActionsController {
             @ApiResponse(code = 500, message = "Internal Server Error")
     })
     public ResponseEntity<BulkCallbackResponse> generateBulkScheduleConfirmation(
-            @RequestBody BulkRequest bulkRequest) {
+            @RequestBody BulkRequest bulkRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
         log.info("GENERATE BULK SCHEDULE CONFIRMATION ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         return ResponseEntity.ok(BulkCallbackResponse.builder()
                 .data(bulkRequest.getCaseDetails().getCaseData())
@@ -472,6 +581,11 @@ public class BulkActionsController {
             @RequestBody BulkRequest bulkRequest,
             @RequestHeader(value = "Authorization") String userToken) {
         log.info("PRE ACCEPT BULK ---> " + LOG_MESSAGE + bulkRequest.getCaseDetails().getCaseId());
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error("Invalid Token {}", userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
 
         List<SubmitEvent> submitEvents = bulkSearchService.retrievalCasesForPreAcceptRequest(bulkRequest.getCaseDetails(), userToken);
